VID-890 make search form initial namespaces configurable
