+++ skins/oasis/modules/SearchController.class.php
@@ -6,6 +6,9 @@
  */
 
 class SearchController extends WikiaController {
+
+	use Wikia\Search\Traits\NamespaceConfigurable;
+
 	public function executeIndex() {
 		$this->setVal('specialSearchUrl', SpecialPage::getTitleFor( 'Search' )->getFullUrl());
 		$this->searchterm = $this->wg->request->getVal('search');
@@ -17,10 +20,18 @@ class SearchController extends WikiaController {
 			$this->noautocomplete = $this->request->getVal( 'noautocomplete' );
 		}
 
-		$this->nonamespaces = $this->wg->request->getVal('nonamespaces');
-		if ( !isset( $this->nonamespaces ) ) {
-			$this->nonamespaces = $this->request->getVal( 'nonamespaces' );
+		$searchParams = [];
+		if (! $this->wg->request->getVal('nonamespaces', $this->request->getVal( 'nonamespaces' ), false ) ) { // why both?
+			$searchConfig = new Wikia\Search\Config();
+			$this->setNamespacesFromRequest($searchConfig, $this->wg->User);
+			foreach ( $searchConfig->getNamespaces() as $namespaceInt ) {
+				$searchParams['ns'.$namespaceInt] = 1;
+			}
 		}
+		if ( $this->wg->CityId == Wikia\Search\QueryService\Select\Dismax\Video::VIDEO_WIKI_ID ) {
+			$searchParams['filters[]'] = 'is_video';
+		}
+		$this->searchParams = $searchParams;
 
 		$this->fulltext = $this->wg->User->getOption('enableGoSearch') ? 0 : 'Search';
 		$this->placeholder = wfMsg('Tooltip-search', $this->wg->Sitename);
