+++ skins/oasis/modules/templates/Search_Index.php
@@ -4,10 +4,9 @@
 <form id="<?= $searchFormId; ?>" class="WikiaSearch<?= empty($noautocomplete) ? '' : ' noautocomplete' ?>" action="<?= $specialSearchUrl; ?>" method="get">
 	<input type="text" name="search" placeholder="<?= $placeholder ?>" autocomplete="off" accesskey="f" value="<?= htmlspecialchars( $searchterm ) ?>">
 	<input type="hidden" name="fulltext" value="<?= $fulltext ?>">
-	<?php if( !$nonamespaces ) : ?>
-		<input type="hidden" name="<?= "ns".NS_MAIN; ?>" value="1">
-		<input type="hidden" name="<?= "ns".NS_CATEGORY; ?>" value="1">
-	<?php endif; ?>
+	<?php foreach( $searchParams as $name => $value ) : ?>
+		<input type="hidden" name="<?= $name ?>" value="<?= $value ?>">
+	<?php endforeach; ?>
 	<input type="submit">
 	<button class="wikia-button"><img src="<?= $wg->BlankImgUrl ?>" class="sprite search" height="17" width="21"></button>
 </form>
